Fix tests on julia 0.5 by depending on Primes.jl
isprime() has been moved to a package in julia 0.5.
We currently use isprime in the tests to reach different code paths
at a non-regular frequency while iterating a loop.  Perhaps it would
be better to figure out a different way to do this (and get rid of
the Primes.jl dependency), but this will fix the tests for now.