Replace control char \u1b[A with \r.
This fixes https://github.com/tpapp/julia-repl/issues/13 using the solution
already employed by `Base.Terminals`, see
https://github.com/JuliaLang/julia/commit/c9f28da19bfe061280fa91120c08ee851e1ff31e

I am not an expert on terminal control characters, but since it is used by Base
it must have seen a lot of testing so I assume it is correct. I tested on Linux,
using various terminals (also with tmux), and ansi-term inside Emacs, and
existing behavior is maintained.

(the neatest solution, of course, would be factoring out the control characters
in Base.Terminals in such a way that they could be sent to any ::IO).