Use stdout's displaysize for size computation
Surprisingly, the argumentless default returns the values of the
environment variables `ROWS` and `COLUMNS`, which (at least on
gnome-terminal) are not necessarily up-to-date with window resizes.
Adding `stdout` gives a size that corresponds to current size.

A followup feature could be to poll this width to stay up-to-date
with window resizes.